moved peer filter parser implementation to own file
